use global context object to initialize heroku client
