+++ index.js
@@ -35,3 +35,4 @@ exports.linewrap = require('./lib/linewrap')
 exports.Spinner = require('./lib/spinner')
 exports.exit = require('./lib/exit').exit
 exports.login = require('./lib/login')
+exports.heroku = require('./lib/heroku')
