+++ lib/command.js
@@ -1,49 +1,7 @@
 'use strict'
+/* globals context */
 
-let Heroku = require('heroku-client')
 let cli = require('..')
-
-function apiMiddleware (response, cb) {
-  let warning = response.headers['x-heroku-warning'] || response.headers['warning-message']
-  if (warning) cli.action.warn(warning)
-  cb()
-}
-
-function heroku (context) {
-  let host = context.apiUrl || 'https://api.heroku.com'
-  let opts = {
-    userAgent: context.version,
-    debug: context.debug,
-    debugHeaders: context.debugHeaders,
-    token: context.auth ? context.auth.password : null,
-    host: host,
-    headers: {},
-    rejectUnauthorized: !(process.env.HEROKU_SSL_VERIFY === 'disable' || host.endsWith('herokudev.com')),
-    middleware: apiMiddleware
-  }
-  if (process.env.HEROKU_HEADERS) {
-    Object.assign(opts.headers, JSON.parse(process.env.HEROKU_HEADERS))
-  }
-  if (context.secondFactor) {
-    Object.assign(opts.headers, {'Heroku-Two-Factor-Code': context.secondFactor})
-  }
-  if (context.reason) {
-    Object.assign(opts.headers, {'X-Heroku-Sudo-Reason': context.reason})
-  }
-  cli.heroku = new Heroku(opts)
-  return cli.heroku
-}
-
-let httpsProxy = process.env.HTTPS_PROXY || process.env.https_proxy || process.env.HTTP_PROXY || process.env.http_proxy
-
-function setupHttpProxy () {
-  const url = require('url')
-  cli.hush(`proxy set to ${httpsProxy}`)
-  let proxy = url.parse(httpsProxy)
-  process.env.HEROKU_HTTP_PROXY_HOST = proxy.hostname
-  process.env.HEROKU_HTTP_PROXY_PORT = proxy.port
-  process.env.HEROKU_HTTP_PROXY_AUTH = proxy.auth
-}
 
 function errHandlerOpts (rollbar, context) {
   rollbar = rollbar || {}
@@ -99,7 +57,7 @@ function twoFactorPrompt (options, context) {
     })
     .then(function (secondFactor) {
       if (options.preauth) {
-        return cli.preauth(context.app, heroku(context), secondFactor)
+        return cli.preauth(context.app, secondFactor)
       } else {
         context.secondFactor = secondFactor
       }
@@ -114,19 +72,18 @@ function reasonPrompt (context) {
 }
 
 module.exports = function command (options, fn) {
-  return function (context) {
+  return function () {
     if (typeof options === 'function') [fn, options] = [options, {}]
-    if (httpsProxy) setupHttpProxy()
     cli.color.enabled = context.supportsColor
     let handleErr = cli.errorHandler(errHandlerOpts(options.rollbar, context))
     let run = function () {
-      let p = fn(context, heroku(context))
+      let p = fn(context, cli.heroku)
       if (!p.catch) return
       return p.catch(function (err) {
         if (err && err.body && err.body.id === 'unauthorized') {
           cli.error(err.body.message || 'Unauthorized')
           relogin()
-            .then(apiKey => { context.auth = {password: apiKey} })
+            .then(apiKey => { cli.heroku.token = context.auth = {password: apiKey} })
             .then(run)
             .catch(handleErr)
         } else if (err && err.body && err.body.id === 'sudo_reason_required') {
