+++ lib/preauth.js
@@ -1,5 +1,7 @@
 'use strict'
-let util = require('./util')
+
+const util = require('./util')
+const heroku = require('./heroku')
 
 /**
  * preauth will make an API call to preauth a user for an app
@@ -11,14 +13,11 @@ let util = require('./util')
  * yubikeys.
  *
  * @param {String} app the app to preauth against
- * @param {Heroku} heroku a heroku api client
  * @param {String} secondFactor a second factor code
  * @return {Promise} A promise fulfilled when the preauth is complete
  */
-function preauth (app, heroku, secondFactor) {
-  return heroku.request({
-    method: 'PUT',
-    path: `/apps/${app}/pre-authorizations`,
+function preauth (app, secondFactor) {
+  return heroku.put(`/apps/${app}/pre-authorizations`, {
     headers: { 'Heroku-Two-Factor-Code': secondFactor }
   })
 }
